refactor classpath to use pluggable strategies
