+++ bundles/com.salesforce.bazel-java-sdk/META-INF/MANIFEST.MF
@@ -21,6 +21,7 @@ Export-Package: com.salesforce.bazel.sdk.aspect,
  com.salesforce.bazel.sdk.init,
  com.salesforce.bazel.sdk.lang.jvm,
  com.salesforce.bazel.sdk.lang.jvm.classpath,
+ com.salesforce.bazel.sdk.lang.jvm.classpath.impl,
  com.salesforce.bazel.sdk.lang.jvm.external,
  com.salesforce.bazel.sdk.logging,
  com.salesforce.bazel.sdk.model,
