+++ bundles/com.salesforce.bazel-java-sdk/src/main/java/com/salesforce/bazel/sdk/lang/jvm/classpath/JvmClasspath.java
@@ -1,13 +1,20 @@
 package com.salesforce.bazel.sdk.lang.jvm.classpath;
 
 /**
- * Interface for generating the classpath for a Bazel package.
+ * Interface for generating the classpath for a Bazel package. An instance of this interface
+ * represents a single package.
  */
 public interface JvmClasspath {
 
     /**
      * Computes the JVM classpath for the associated Bazel package
      */
-    BazelJvmClasspathResponse getClasspathEntries();
+    JvmClasspathResponse getClasspathEntries();
+
+    /**
+     * Requests the classpath to clear all cached state such that the next call to getClasspathEntries()
+     * will rebuild internal state. 
+     */
+    public void clean();
 
 }
