+++ bundles/com.salesforce.bazel.eclipse.common/src/main/java/com/salesforce/bazel/eclipse/classpath/BazelGlobalSearchClasspathContainer.java
@@ -53,7 +53,7 @@ import com.salesforce.bazel.eclipse.runtime.api.ResourceHelper;
 import com.salesforce.bazel.eclipse.runtime.impl.EclipseWorkProgressMonitor;
 import com.salesforce.bazel.sdk.command.BazelCommandLineToolConfigurationException;
 import com.salesforce.bazel.sdk.index.jvm.BazelJvmIndexClasspath;
-import com.salesforce.bazel.sdk.lang.jvm.classpath.BazelJvmClasspathResponse;
+import com.salesforce.bazel.sdk.lang.jvm.classpath.JvmClasspathResponse;
 import com.salesforce.bazel.sdk.lang.jvm.external.BazelExternalJarRuleManager;
 import com.salesforce.bazel.sdk.logging.LogHelper;
 import com.salesforce.bazel.sdk.model.BazelConfigurationManager;
@@ -113,11 +113,11 @@ public class BazelGlobalSearchClasspathContainer extends BaseBazelClasspathConta
     }
 
     @Override
-    public BazelJvmClasspathResponse computeClasspath() {
+    public JvmClasspathResponse computeClasspath() {
         BazelWorkspace bazelWorkspace = ComponentContext.getInstance().getBazelWorkspace();
         if (Objects.isNull(config) || !config.isGlobalClasspathSearchEnabled() || (bazelWorkspace == null)) {
             // user has disabled the global search feature, or hasnt imported a bazel workspace yet
-            return new BazelJvmClasspathResponse();
+            return new JvmClasspathResponse();
         }
 
         if (!bazelWorkspace.getName().equals(bazelWorkspaceName)) {
@@ -135,7 +135,7 @@ public class BazelGlobalSearchClasspathContainer extends BaseBazelClasspathConta
         // the Java SDK will produce a list of logical classpath entries
         long startTime = System.currentTimeMillis();
         EclipseWorkProgressMonitor monitor = new EclipseWorkProgressMonitor(null);
-        BazelJvmClasspathResponse computedClasspath = bazelJvmIndexClasspath.getClasspathEntries(monitor);
+        JvmClasspathResponse computedClasspath = bazelJvmIndexClasspath.getClasspathEntries(monitor);
         long endTime = System.currentTimeMillis();
 
         LOG.info("completed indexing in [{}] milliseconds.", (endTime - startTime));
